CRIS: Fix PR middle-end/113109; "throw" failing
No test-case, but the regress-367 from r14-6674-g4759383245ac97 is
"back" to regress-10 for cris-elf+cris-sim with this patch applied
to gcc from that revision.

Also, I wonder why none of those other targets with a MEM for
EH_RETURN_HANDLER_RTX with an address relative to
frame_pointer_rtx (as opposed to hard_frame_pointer_rtx or
virtual_incoming_args_rtx) don't see the same problem.

Oh well.  Merry Xmas.

brgds, H-P

-- >8 --
TL;DR: the "dse1" pass removed the eh-return-address store.  The
PA also marks its EH_RETURN_HANDLER_RTX as volatile, for the same
reason, as does visum.  See PR32769 - it's the same thing on PA.

Conceptually, it's logical that stores to incoming args are
optimized out on the return path or if no loads are seen -
at least before epilogue expansion, when the subsequent load
isn't seen in the RTL, as is the case for the "dse1" pass.

I haven't looked into why this problem, that appeared for the PA
already in 2007, was seen for CRIS only recently (with
r14-6674-g4759383245ac97).

	PR middle-end/113109
	* config/cris/cris.cc (cris_eh_return_handler_rtx): New function.
	* config/cris/cris-protos.h (cris_eh_return_handler_rtx): Prototype.
	* config/cris/cris.h (EH_RETURN_HANDLER_RTX): Redefine to call
	cris_eh_return_handler_rtx.