libstdc++ testsuite/20_util/hash/quality.cc: Increase timeout 3x
Tested for mmix and observing the increased timeout in the .log
file - and the test passing.

Ok to commit?  Or better suggestions?

-- >8 --
Testing for mmix (a 64-bit target using Knuth's simulator).  The test
is largely pruned for simulators, but still needs 5m57s on my laptop
from 3.5 years ago to run to successful completion.  Perhaps slow
hosted targets could also have problems so increasing the timeout
limit, not just for simulators but for everyone, and by more than a
factor 2.

	* testsuite/20_util/hash/quality.cc: Increase timeout by a factor 3.