Add -falign-all-functions
Hi,
this patch adds new option -falign-all-functions which works like
-falign-functions, but applies to all functions including those in cold
regions.  As discussed in the PR log, this is needed for atomically
patching function entries in the kernel.

An option would be to make -falign-function mandatory, but I think it is not a
good idea, since original purpose of -falign-funtions is optimization of
instruction decode and cache size.  Having -falign-all-functions is
backwards compatible.  Richi also suggested extending syntax of the
-falign-functions parameters (which is already non-trivial) but it seems
to me that having separate flag is more readable.

Bootstrapped/regtested x86_64-linux, OK for master and later
backports to release branches?

gcc/ChangeLog:

	PR middle-end/88345
	* common.opt: Add -falign-all-functions
	* doc/invoke.texi: Add -falign-all-functions.
	(-falign-functions, -falign-labels, -falign-loops): Document
	that alignment is ignored in cold code.
	* flags.h (align_loops): Reindent.
	(align_jumps): Reindent.
	(align_labels): Reindent.
	(align_functions): Reindent.
	(align_all_functions): New macro.
	* opts.cc (common_handle_option): Handle -falign-all-functions.
	* toplev.cc (parse_alignment_opts): Likewise.
	* varasm.cc (assemble_start_function): Likewise.