c++: reference variable as default targ [PR101463]
Bootstrapped and regtested on x86_64-pc-linux-gnu, does this
look OK for trunk?

-- >8 --

Here during default template argument substitution we wrongly consider
the (substituted) default arguments v and vt<int> as value-dependent[1]
which ultimately leads to deduction failure for the calls.

The bogus value_dependent_expression_p result aside, I noticed
type_unification_real during default targ substitution keeps track of
whether all previous targs are known and non-dependent, as is the case
for these calls.  And in such cases it should be safe to avoid checking
dependence of the substituted default targ and just assume it's not.
This patch implements this optimization, which lets us accept both
testcases by sidestepping the value_dependent_expression_p issue
altogether.

[1]: The reason we consider these reference variables value-dependent is
due to a workaround in value_dependent_expression_p:

  case VAR_DECL:
    ...
    else if (TYPE_REF_P (TREE_TYPE (expression)))
      /* FIXME cp_finish_decl doesn't fold reference initializers.  */
      return true;
    ...

added by r5-5022-g51d72abe5ea04e.  I'm not sure if this workaround
is needed anymore, but naively removing it seems safe as far as
bootstrap+regtest is concerned (the only change is that we issue more
-Wmissing-braces warnings ahead of time in cpp0x/initlist123.C), and
lets us accept the first testcase.

Unfortunately we still reject the second testcase (in which v and vt are
additionally constexpr) for the same reason (bogus value dependence) due
to the subsequent check in v_d_e_p:

      ...
      /* We have a constexpr variable and we're processing a template.  When
         there's lifetime extension involved (for which finish_compound_literal
         used to create a temporary), we'll not be able to evaluate the
         variable until instantiating, so pretend it's value-dependent.  */
      else if (DECL_DECLARED_CONSTEXPR_P (expression)
               && !TREE_CONSTANT (expression))
        return true;

And TREE_CONSTANT isn't set for v and vt<int> because of a workaround in
cp_finish_decl:

          if (decl_maybe_constant_var_p (decl)
              /* FIXME setting TREE_CONSTANT on refs breaks the back end.  */
              && !TYPE_REF_P (type))
            TREE_CONSTANT (decl) = true;

Naively removing this workaround lets us accept the second testcase, but
it re-introduces an ICE in g++.dg/opt/pr78373.C.

	PR c++/101463

gcc/cp/ChangeLog:

	* pt.cc (type_unification_real): Avoid checking dependence of
	a substituted default template argument if we can assume it's
	non-dependent.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1z/nontype6.C: New test.
	* g++.dg/cpp1z/nontype6a.C: New test.