Add __cow_string C string constructor
Hi

While working on the patch to use the cxx11 abi in gnu version namespace
mode I got a small problem with this missing constructor. I'm not sure
that the main patch will be integrated in gcc 14 so I think it is better
if I propose this patch independently.

     libstdc++: Add __cow_string constructor from C string

     The __cow_string is instantiated from a C string in
cow-stdexcept.cc. At the moment
     the constructor from std::string is being used with the drawback of
an intermediate
     potential allocation/deallocation and copy. With the C string
constructor we bypass
     all those operations.

     libstdc++-v3/ChangeLog:

             * include/std/stdexcept (__cow_string(const char*)): New
definition.
             * src/c++11/cow-stdexcept.cc (__cow_string(const char*)):
New definition and
             declaration.

Tested under Linux x64, ok to commit ?

François