btf: print string position as comment for validation and testing purposes.
Hi everyone,

This patch adds a comment to the BTF strings regarding their position
within the section. This is useful for assembly inspection purposes.

Regards,
Cupertino

When using -dA, this function was only printing as comment btf_string or
btf_aux_string.
This patch changes the comment to also include the position of the
string within the section in hexadecimal format.

gcc/ChangeLog:
	* btfout.cc (output_btf_strs): Changed.