Fix PR rtl-optimization/113140
This is a small regression present on the mainline and 13 branch, although the
underlying problem has probably been there for ages, in the form of a segfault
during the delay slot scheduling pass, for a function that falls through to
exit without any RTL instruction generated for the end of function.

Tested on SPARC64/Linux, applied on the mainline and 13 branch.

2024-01-09  Eric Botcazou  <ebotcazou@adacore.com>

	PR rtl-optimization/113140
	* reorg.cc (fill_slots_from_thread): If we are to branch after the
	last instruction of the function, create an end label.

2024-01-09  Eric Botcazou  <ebotcazou@adacore.com>

	* g++.dg/opt/delay-slot-2.C: New test.