Fix PR ada/113195
This is a small regression present on the mainline and 13 branch, in the form
of an internal error in gigi on anonymous access type equality.  We now need
to also accept them too for anonymous access types that point to compatible
object subtypes in the language sense.

Tested on SPARC64/Linux, applied on the mainline and 13 branch.

2024-01-09  Eric Botcazou  <ebotcazou@adacore.com>

	PR ada/113195
	* gcc-interface/utils2.cc (build_binary_op) <EQ_EXPR>: Relax
	assertion for regular pointer types.

2024-01-09  Eric Botcazou  <ebotcazou@adacore.com>

	* gnat.dg/specs/anon4.ads: New test.