ada: Avoid xref on out params of TSS
For an actual passed as an 'in out' parameter of a type support
subprogram such as deep finalize, do not count it as a read
reference of the actual. Clearly these should not count.
Furthermore, counting them causes different warnings in -gnatc
mode compared to normal mode, because the calls only exist in
normal mode, which would disable the warnings. Such warnings now
occur in both modes, instead of just with -gnatc.

gcc/ada/

	* lib-xref.adb (Generate_Reference): Do not count it as a read
	reference if we're calling a TSS.

Tested on x86_64-pc-linux-gnu, committed on master.