Fix minor bug on mn103 port
Richard Sandiford debugged a failure on the mn103 port with his
late-combine patches down to the subdi3 pattern not specifying the isa
on alternatives which required newer variants of the chip family.

This patch adds the missing isa attribute and the port now works with
his late-combine patch.  I'm pushing this to the trunk on his behalf.

Jeff
commit 9aaed2c1d7e54cee6966c632ced80e643525fe89
Author: Richard Sandiford <richard.sandiford@arm.com>
Date:   Tue Jan 9 10:07:09 2024 -0700

    [committed] Fix minor bug on mn103 port

    Richard Sandiford debugged a failure on the mn103 port with his late-combine
    patches down to the subdi3 pattern not specifying the isa on alternatives which
    required newer variants of the chip family.

    This patch adds the missing isa attribute and the port now works with his
    late-combine patch.  I'm pushing this to the trunk on his behalf.

    gcc/
            * config/mn10300/mn10300.md (subdi3_degenerate): Add isa attribute.